+++ skellytracker/trackers/rtmpose_tracker/rtmpose_detector.py
@@ -12,6 +12,7 @@ import onnxruntime
 from numpy.typing import NDArray
 
 from skellytracker.trackers.base_tracker.base_tracker_abcs import BaseDetector, BaseDetectorConfig, TrackerType
+from skellytracker.trackers.rtmpose_tracker.rtmpose_detector_config import RTMPoseDetectorConfig
 from skellytracker.trackers.rtmpose_tracker.rtmpose_observation import RTMPoseObservation
 from skellytracker.trackers.rtmpose_tracker.rtmpose_session import (
     ExecutionProviderName,
@@ -110,45 +111,6 @@ def _verify_ort_install_sane() -> None:
         raise RuntimeError(
             "ONNX Runtime reports no execution providers. The install is broken."
         )
-
-
-# Backwards-compatible alias maintained for existing callers / configs that
-# still pass `device="cuda"`. New code should use `execution_provider`.
-_DEVICE_TO_PROVIDER: dict[str, ExecutionProviderName] = {
-    "cuda": "cuda",
-    "trt": "trt",
-    "tensorrt": "trt",
-    "mps": "coreml",
-    "coreml": "coreml",
-    "cpu": "cpu",
-}
-
-
-def _default_execution_provider() -> ExecutionProviderName:
-    import sys
-    return "coreml" if sys.platform == "darwin" else "cuda"
-
-
-class RTMPoseDetectorConfig(BaseDetectorConfig):
-    tracker_type: Literal[TrackerType.RTMPOSE] = TrackerType.RTMPOSE
-    confidence_threshold: float = 0.5
-    mode: str = "performance"
-    backend: str = "onnxruntime"
-    device: str = "auto"
-    # When set, takes precedence over `device`. Drives the actual ORT provider selection.
-    execution_provider: ExecutionProviderName | None = None
-    # Which GPU to use. None = auto-select the device with the most VRAM at session creation.
-    device_id: int | None = None
-    # Keep only the N highest-confidence YOLOX detections. None = keep all.
-    # Set to 1 for single-person use to suppress false positives from background clutter.
-    max_persons: int | None = None
-
-    def resolved_provider(self) -> ExecutionProviderName:
-        if self.execution_provider is not None:
-            return self.execution_provider
-        if self.device in _DEVICE_TO_PROVIDER:
-            return _DEVICE_TO_PROVIDER[self.device]
-        return _default_execution_provider()
 
 
 @dataclass
